refactor: 提取剪贴板/配置/工具方法，完成 FmMain.cs 拆分
- FmMain.Clipboard.cs: 剪贴板监听与处理 (9 methods)
- FmMain.Config.cs: 配置加载与初始化 (9 methods)
- FmMain.Utilities.cs: 热键/搜索/文件/杂项 (8 methods)
- FmMain.cs 现仅保留字段、构造函数、WndProc、Form_Resize 等核心
