+++ FmMain.cs
@@ -237,31 +237,6 @@ namespace TrOCR
 			this.RichBoxBody_T.TemporaryTranslateRequested += RichBoxBody_T_OnTemporaryTranslateRequested;
 	
         }
-        // FmMain.cs
-
-        // 这是一个专门用来“拦截并报错”的事件
-        private void ShowConfigWarning_MouseDown(object sender, MouseEventArgs e)
-        {
-            if (e.Button != MouseButtons.Left) return;
-
-            // 判断是哪个菜单触发的，显示对应的提示
-            string msg = "检测到您尚未配置 AI 接口。\n\n请先去设置里配置";
-            if (sender == ai_menu_trans) msg = "检测到您尚未配置 AI 翻译接口。\n\n请先去设置里配置";
-
-            MessageBox.Show(msg, "配置提示", MessageBoxButtons.OK, MessageBoxIcon.None);
-
-        }
-
-        /// <summary>
-        /// 点击加载按钮时触发的事件处理函数，将窗体最小化并隐藏
-        /// </summary>
-        /// <param name="sender">事件发送者</param>
-        /// <param name="e">事件参数</param>
-        private void Load_Click(object sender, EventArgs e)
-		{
-			WindowState = FormWindowState.Minimized;
-			Visible = false;
-		}
 		/// <summary>
 		/// 从配置文件加载并应用窗口的基础尺寸
 		/// </summary>
@@ -310,24 +285,6 @@ namespace TrOCR
             IniHelper.SetValue("WindowState", "Width", logicalWidth.ToString());
             IniHelper.SetValue("WindowState", "Height", logicalHeight.ToString());
         }
-
-		// 【新增】临时翻译事件的处理器
-		private void RichBoxBody_T_OnTemporaryTranslateRequested(object sender, TempTranslateEventArgs e)
-		{
-					// 【新增调试代码】弹窗显示接收到的语言代码
-					// MessageBox.Show($"接收到临时翻译请求：\n源语言: {e.SourceLanguage}\n目标语言: {e.TargetLanguage}");
-					// 确保翻译的文本是最新的
-					typeset_txt = RichBoxBody.Text;
-
-			if (string.IsNullOrWhiteSpace(typeset_txt))
-			{
-				MessageBox.Show("请输入需要翻译的文本！");
-				return;
-			}
-
-			// 调用我们改造后的翻译方法，并传入临时的语言代码
-			trans_Calculate(e.SourceLanguage, e.TargetLanguage);
-		}
         /// <summary>
         /// 重写Windows窗体的消息处理方法，用于处理发送到窗口的各种消息
         /// </summary>
@@ -763,276 +720,7 @@ namespace TrOCR
 		    }
 		}
 
-        /// <summary>
-        /// 停止所有正在运行的逻辑定时器 (用于重置状态、关闭窗口或切换模式时)，暂时不使用.手动处理了
-		/// 最核心的原则是：只要发生了“模式切换”或“窗口状态改变”，都应该注意是否停止 translationTimer等定时器。
-        /// </summary>
-        private void StopAllActiveTimers()
-        {
-            if (translationTimer != null)
-                translationTimer.Stop();
-
-            if (clipboardDebounceTimer != null)
-                clipboardDebounceTimer.Stop();
-
-            // autoCopyLockTimer 一般不需要强制停止，因为它负责释放锁，强制停止可能导致锁死。
-            // trayClickTimer 处理托盘点击，也不建议随意停止。
-        }
-
 		#endregion
-// ====================================================================================================================
-		// **主菜单事件**
-		//
-		// 处理主文本框（RichBoxBody）的右键上下文菜单事件。
-		// - MainCopyClick(): 实现“复制”功能。
-		// - Main_SelectAll_Click(): 实现“全选”功能。
-		// - Main_paste_Click(): 实现“粘贴”功能。
-		// ====================================================================================================================
-		#region 主菜单事件
-		/// <summary>
-		/// 复制操作事件处理函数
-		/// </summary>
-		/// <param name="sender">事件发送者</param>
-		/// <param name="e">事件参数</param>
-		private void MainCopyClick(object sender, EventArgs e)
-		{
-			RichBoxBody.Focus();
-			RichBoxBody.richTextBox1.Copy();
-		}
-
-		/// <summary>
-		/// 全选操作事件处理函数
-		/// </summary>
-		/// <param name="sender">事件发送者</param>
-		/// <param name="e">事件参数</param>
-		private void Main_SelectAll_Click(object sender, EventArgs e)
-		{
-			RichBoxBody.Focus();
-			RichBoxBody.richTextBox1.SelectAll();
-		}
-
-		/// <summary>
-		/// 粘贴操作事件处理函数
-		/// </summary>
-		/// <param name="sender">事件发送者</param>
-		/// <param name="e">事件参数</param>
-		private void Main_paste_Click(object sender, EventArgs e)
-		{
-			RichBoxBody.Focus();
-			RichBoxBody.richTextBox1.Paste();
-		}
-		#endregion
-// ====================================================================================================================
-		// **OCR 引擎调用**
-		//
-		// 包含调用不同 OCR 服务（腾讯、微信、白描、百度等）的实现方法。
-		// - OCR_Tencent(): 调用腾讯云 OCR API（通用版与高精度版）进行文字识别。
-		// - OCR_WeChat(): 调用微信 OCR API 进行文字识别。
-		// - OCR_Baimiao(): 调用白描 OCR API 进行文字识别。
-		// - OCR_baidu(), OCR_baidu_accurate(): 调用百度标准版和高精度版OCR API。
-		// - OCR_youdao(): 调用有道 OCR API 进行文字识别。
-		// ====================================================================================================================
-// ====================================================================================================================
-		// **OCR 接口切换 (事件)**
-		//
-		// 包含用户在界面上选择不同 OCR 引擎的事件处理程序。
-		// 每个事件处理程序通过调用 OCR_foreach(string name) 方法来更新当前使用的 OCR 接口。
-		// ====================================================================================================================
-// ====================================================================================================================
-		// **文本操作与格式化**
-		//
-		// 提供中英文标点符号转换等文本处理功能。
-		// - change_Chinese_Click(): 将文本中的英文标点符号转换为中文标点。
-		// - change_English_Click(): 将文本中的中文标点符号转换为英文标点。
-		// - punctuation_ch_en(): 具体的中文转英文标点实现。
-		// ====================================================================================================================
-// ====================================================================================================================
-		// **配置文件与初始化**
-		//
-		// 负责加载和保存应用程序的配置信息（config.ini）。
-		// - saveIniFile(): 保存当前配置到 ini 文件。
-		// - LoadTranslateConfig(): 从 ini 文件加载所有翻译服务的配置（源语言、目标语言、密钥等）。
-		// - InitConfig(): 在程序启动时初始化所有配置，包括 OCR 接口、翻译接口、热键和各 API 密钥。
-		// - tray_Set_Click(): 处理托盘菜单中的“设置”点击事件，打开设置窗口并重新加载所有配置。
-		// ====================================================================================================================
-		#region 配置文件与初始化
-		/// <summary>
-		/// 保存当前选择的OCR接口配置到配置文件中
-		/// </summary>
-		public void saveIniFile()
-		{
-			IniHelper.SetValue("配置", "接口", interface_flag);
-		}
-
-
-		/// <summary>
-		/// 安全获取配置值的辅助方法
-		/// </summary>
-		/// <param name="section">配置节</param>
-		/// <param name="key">配置键</param>
-		/// <param name="defaultValue">默认值</param>
-		/// <returns>配置值或默认值</returns>
-		private string GetConfigValueSafely(string section, string key, string defaultValue = "")
-		{
-			try
-			{
-				var value = IniHelper.GetValue(section, key);
-				return (string.IsNullOrEmpty(value) || value == "发生错误") ? defaultValue : value;
-			}
-			catch (Exception ex)
-			{
-				System.Diagnostics.Debug.WriteLine($"读取配置失败 [{section}][{key}]: {ex.Message}");
-				return defaultValue;
-			}
-		}
-
-		/// <summary>
-		/// 读取配置文件(config.ini)，按照配置文件初始化应用程序配置到内存，包括OCR接口、翻译接口、快捷键和各种API密钥
-		/// 读取配置文件，并根据其内容将应用程序的各项配置初始化到内存中。这包括设置OCR与翻译接口、注册全局快捷键以及加载各种API密钥。
-		/// </summary>
-		private void InitConfig()
-		{
-			// 初始化API菜单
-			InitializeApiMenus();
-			// 【新增】加载 AI 动态菜单
-			LoadCustomOpenAIMenus();
-			LoadCustomOpenAITransMenus();
-			
-			// 初始化OCR接口配置
-			interface_flag = GetConfigValueSafely("配置", "接口", "搜狗");
-			if (string.IsNullOrEmpty(interface_flag))
-			{
-				interface_flag = "搜狗";
-				IniHelper.SetValue("配置", "接口", interface_flag);
-			}
-			OCR_foreach(interface_flag);
-			
-			// 初始化翻译接口配置
-			StaticValue.Translate_Current_API = GetConfigValueSafely("配置", "翻译接口", "Bing2");
-			if (string.IsNullOrEmpty(StaticValue.Translate_Current_API))
-			{
-				StaticValue.Translate_Current_API = "Bing2";
-			}
-			Trans_foreach(StaticValue.Translate_Current_API);
-			LoadTranslateConfig();
-			
-			// 初始化快捷键配置
-			var filePath = AppDomain.CurrentDomain.BaseDirectory + "Data\\config.ini";
-			
-			// 安全加载热键配置的辅助方法
-			Action<string, string, int> loadHotkey = (section, key, flagId) =>
-			{
-				try
-				{
-					var hotkeyValue = IniHelper.GetValue(section, key);
-					if (!string.IsNullOrEmpty(hotkeyValue) && 
-					    hotkeyValue != "请按下快捷键" && 
-					    hotkeyValue != "发生错误")
-					{
-						SetHotkey("None", "", hotkeyValue, flagId);
-					}
-				}
-				catch (Exception ex)
-				{
-					// 记录错误但不中断程序执行
-					System.Diagnostics.Debug.WriteLine($"加载热键配置失败 [{section}][{key}]: {ex.Message}");
-				}
-			};
-
-			// 加载各个热键配置
-			loadHotkey("快捷键", "文字识别", 200);
-			loadHotkey("快捷键", "翻译文本", 205);
-			loadHotkey("快捷键", "记录界面", 206);
-			loadHotkey("快捷键", "识别界面", 235);
-			loadHotkey("快捷键", "输入翻译", 240);
-			loadHotkey("快捷键", "静默识别", 250);
-			loadHotkey("快捷键", "截图翻译", 260);
-
-			// --- 加载OCR密钥 ---
-			// 加载百度OCR密钥
-			StaticValue.BD_API_ID = TrOCRUtils.LoadSetting("密钥_百度", "secret_id", "");
-			StaticValue.BD_API_KEY = TrOCRUtils.LoadSetting("密钥_百度", "secret_key","");
-			StaticValue.BD_LANGUAGE = TrOCRUtils.LoadSetting("密钥_百度", "language_code","CHN_ENG");
-
-			// 加载百度表格识别密钥
-			StaticValue.BD_TABLE_API_ID = TrOCRUtils.LoadSetting("密钥_百度表格", "secret_id","");
-			StaticValue.BD_TABLE_API_KEY = TrOCRUtils.LoadSetting("密钥_百度表格", "secret_key","");
-			
-			// 【新增】加载百度手写识别密钥
-			StaticValue.BD_HANDWRITING_API_ID = TrOCRUtils.LoadSetting("密钥_百度手写", "secret_id", "");
-			StaticValue.BD_HANDWRITING_API_KEY = TrOCRUtils.LoadSetting("密钥_百度手写", "secret_key", "");
-			StaticValue.BD_HANDWRITING_LANGUAGE = TrOCRUtils.LoadSetting("密钥_百度手写", "language_code", "CHN_ENG");
-
-			// 加载腾讯OCR密钥
-			StaticValue.TX_API_ID = TrOCRUtils.LoadSetting("密钥_腾讯", "secret_id","");		
-			StaticValue.TX_API_KEY = TrOCRUtils.LoadSetting("密钥_腾讯", "secret_key","");	
-			StaticValue.TX_LANGUAGE = TrOCRUtils.LoadSetting("密钥_腾讯", "language_code","zh");
-
-			// 加载腾讯高精度OCR密钥
-			StaticValue.TX_ACCURATE_API_ID = TrOCRUtils.LoadSetting("密钥_腾讯高精度", "secret_id","");		
-			StaticValue.TX_ACCURATE_API_KEY = TrOCRUtils.LoadSetting("密钥_腾讯高精度", "secret_key","");		
-			StaticValue.TX_ACCURATE_LANGUAGE = TrOCRUtils.LoadSetting("密钥_腾讯高精度", "language","auto");
-			
-			// 加载腾讯表格v3的OCR密钥
-			StaticValue.TX_TABLE_API_ID = TrOCRUtils.LoadSetting("密钥_腾讯表格v3", "secret_id","");			
-			StaticValue.TX_TABLE_API_KEY = TrOCRUtils.LoadSetting("密钥_腾讯表格v3", "secret_key","");			
-
-			// 加载百度高精度OCR密钥
-			StaticValue.BD_ACCURATE_API_ID = TrOCRUtils.LoadSetting("密钥_百度高精度", "secret_id","");			
-			StaticValue.BD_ACCURATE_API_KEY = TrOCRUtils.LoadSetting("密钥_百度高精度", "secret_key","");			
-			StaticValue.BD_ACCURATE_LANGUAGE = TrOCRUtils.LoadSetting("密钥_百度高精度", "language_code","CHN_ENG");			
-	
-			// --- 加载白描OCR凭据 ---
-			StaticValue.BaimiaoUsername = TrOCRUtils.LoadSetting("密钥_白描", "username","");
-			StaticValue.BaimiaoPassword = TrOCRUtils.LoadSetting("密钥_白描", "password","");
-
-			// 加载持久化的token信息
-			string savedToken = IniHelper.GetValue("密钥_白描", "token");
-			string savedExpiry = IniHelper.GetValue("密钥_白描", "token_expiry");
-			string savedUsername = IniHelper.GetValue("密钥_白描", "token_username");
-			string savedUuid = IniHelper.GetValue("密钥_白描", "device_uuid");
-
-			if (!string.IsNullOrEmpty(savedToken) && savedToken != "发生错误" &&
-			    !string.IsNullOrEmpty(savedUsername) && savedUsername != "发生错误" &&
-			    savedUsername == StaticValue.BaimiaoUsername && // 确保token属于当前用户
-			    DateTime.TryParse(savedExpiry, out DateTime expiry) && DateTime.Now < expiry)
-			{
-			 StaticValue.BaimiaoToken = savedToken;
-			 StaticValue.BaimiaoTokenExpiry = expiry;
-			}
-
-			if (!string.IsNullOrEmpty(savedUuid) && savedUuid != "发生错误")
-			{
-			 StaticValue.BaimiaoDeviceUuid = savedUuid;
-			}
-		}
-		
-		/// <summary>
-		/// 检查并替换文本中的中文冒号为英文冒号
-		/// 当中文冒号前后都是英文字符或标点符号时，将其替换为英文冒号
-		/// </summary>
-		/// <param name="text">需要处理的文本</param>
-		/// <returns>处理后的文本，其中符合条件的中文冒号已被替换为英文冒号</returns>
-		public static string check_ch_en(string text)
-		{
-			var array = text.ToCharArray();
-			for (var i = 0; i < array.Length; i++)
-			{
-				var num = "：".IndexOf(array[i]);
-				if (num != -1 && i - 1 >= 0 && i + 1 < array.Length && contain_en(array[i - 1].ToString()) && contain_en(array[i + 1].ToString()))
-				{
-					array[i] = ":"[num];
-				}
-				if (num != -1 && i - 1 >= 0 && i + 1 < array.Length && contain_en(array[i - 1].ToString()) && contain_punctuation(array[i + 1].ToString()))
-				{
-					array[i] = ":"[num];
-				}
-			}
-			return new string(array);
-		}
-
-
-        #endregion
 
         // ====================================================================================================================
         // **翻译功能**
@@ -1150,267 +838,8 @@ namespace TrOCR
 			LogState("Form_Resize End"); // <--- 添加这一行
 		}
 
-
-        /// <summary>
-        /// 使用“限时状态锁”安全地将数据对象设置到剪贴板，以防止无限循环。
-        /// </summary>
-        /// <param name="data">要复制到剪贴板的对象 (可以是 string, DataObject, Image 等)。</param>
-        public void SetClipboardWithLock(object data)
-        {
-            // 检查传入的对象是否为 null
-            if (data == null) return;
-
-            // 如果传入的是字符串，额外检查它是否为空或仅包含空白字符
-            if (data is string textData && string.IsNullOrWhiteSpace(textData))
-            {
-                return;
-            }
-
-            try
-            {
-                // 1. 激活状态锁
-                isAutoCopying = true;
-                Debug.WriteLine("--- 剪贴板锁已激活 ---");
-
-                // 2. 执行复制操作
-                // 这个调用现在是通用的，可以处理 string、DataObject 等多种类型
-                Clipboard.SetDataObject(data, true, 5, 100);
-
-                // 3. 启动定时器，创建静默期
-                autoCopyLockTimer.Stop(); // 确保定时器被重置
-                autoCopyLockTimer.Start();
-            }
-            catch (Exception ex)
-            {
-                System.Diagnostics.Debug.WriteLine($"带锁设置剪贴板失败: {ex.Message}");
-                // 如果复制失败，必须立即解除锁定
-                isAutoCopying = false;
-                autoCopyLockTimer.Stop();
-            }
-        }
-        /// <summary>
-        /// 使用“限时状态锁”安全地为特定格式设置剪贴板数据。
-        /// 主要用于处理“快速翻译”后的粘贴操作。
-        /// </summary>
-        /// <param name="format">剪贴板格式 (例如 DataFormats.UnicodeText)。</param>
-        /// <param name="data">要为该格式复制的数据。</param>
-        private void SetClipboardDataWithLock(string format, object data)
-        {
-            if (string.IsNullOrEmpty(format) || data == null) return;
-
-            try
-            {
-                isAutoCopying = true;
-                Debug.WriteLine("--- 剪贴板锁已激活 (Data) ---");
-                // 此调用处理 Clipboard.SetData 的情况
-                Clipboard.SetData(format, data);
-                autoCopyLockTimer.Stop();
-                autoCopyLockTimer.Start();
-            }
-            catch (Exception ex)
-            {
-                System.Diagnostics.Debug.WriteLine($"带锁设置剪贴板数据失败: {ex.Message}");
-                isAutoCopying = false;
-                autoCopyLockTimer.Stop();
-            }
-        }
-        /// <summary>
-        /// 原始文本框内容改变事件，用于实现编辑后自动翻译
-        /// </summary>
-        private void RichBoxBody_TextChanged(object sender, EventArgs e)
-		{
-            Debug.WriteLine("启动了RichBoxBody_TextChanged");
-            // 【新增】如果正在流式输出，直接忽略，不启动定时器
-            if (isStreaming) return;
-
-            // 【新增】拦截逻辑
-            // 如果 RichBoxBody 正在进行字体切换，直接返回，不启动翻译倒计时
-            if (RichBoxBody.IsFontChanging)
-            {
-                // 可选：打印日志确认拦截成功
-                 System.Diagnostics.Debug.WriteLine("字体切换中，忽略 TextChanged 事件");
-                return;
-            }
-            // === 【修改步骤 1】读取原始配置字符串 ===
-            string rawConfig = StaticValue.TextChangeAutotranslateDelayRaw;
-            // === 【修改步骤 2】调用解析方法 ===
-            // CheckTextChangeAutoTranslateConfig 会处理：
-            // 1. 解析时间，如果非数字或<=0，返回 false
-            // 2. 解析黑白名单，判断当前接口是否允许
-            // 3. 将解析出的时间赋值给 validDelay
-            if (!CheckTextChangeAutoTranslateConfig(rawConfig, StaticValue.Translate_Current_API, out int validDelay))
-            {
-                // 如果不允许（时间为0、格式错误、或当前接口被排除），直接退出
-                return;
-            }
-
-            // --- 日志: 事件触发入口 ---
-            // 为了日志清晰，将换行符替换为可见的转义字符
-            string currentTextForLog = RichBoxBody.Text.Replace("\r", "\\r").Replace("\n", "\\n");
-			Debug.WriteLine($"---> TextChanged 事件触发。文本: \"{currentTextForLog}\" | isContentFromOcr: {isContentFromOcr} | transtalate_fla: {transtalate_fla}");
-			// 关键修复：添加一个“守卫”，如果文本是默认占位符，则直接忽略，不执行任何逻辑。
-			// 这一步不做也行，因为下面2880行做了事件临时解绑。
-			//这一步做了有个小问题：当OCR识别的结果恰好是 ***该区域未发现文本*** 时，也会直接return,需要优化
-    		// if (RichBoxBody.Text == "***该区域未发现文本***")
-    		// {
-    		//     return;
-    		// }
-
-			// 使用安全的字符串比较方式，避免因 "发生错误" 或空值导致异常
-			bool autoTranslateInputEnabled = StaticValue.InputTranslateAutoTranslate;
-			// 定义纯手动输入状态：非OCR 也 非剪贴板
-        	bool isPureManualInput = !isContentFromOcr && !isFromClipboardListener;
-
-			 // 场景1: 当翻译窗口关闭时 (只处理纯手动输入的情况)
-    		if (transtalate_fla == "关闭")
-    		{
-				Debug.WriteLine("    |--> 满足 [场景1：输入翻译 & 翻译窗口关闭]");
-        		if (isPureManualInput && autoTranslateInputEnabled && !string.IsNullOrWhiteSpace(RichBoxBody.Text))
-				{
-					Debug.WriteLine("        |--> 文本不为空，准备调用 TransClick() 来打开翻译窗口...");
-					translationTimer.Stop();
-                    // 【修改后新增】在此处动态更新 Timer 的间隔
-                    // 这样可以确保：
-                    // 1. 此时 currentDelay 肯定 > 0，赋值安全，不会崩溃
-                    // 2. 如果用户在设置里修改了时间，这里会立即生效，无需重启
-                    // === 【修改步骤 3】使用解析出的有效时间 ===
-                    translationTimer.Interval = validDelay;
-                    translationTimer.Start();
-				}
-				Debug.WriteLine("    |<-- 场景1 结束,定时器已开始或重置。");
-				Debug.WriteLine("---> TextChanged 事件结束。");
-    		    return;
-    		}
-
-    		// 场景2: 当翻译窗口已经打开时
-    		if (transtalate_fla == "开启")
-    		{
-				// 允许重新翻译的条件是：
-				// 1. 内容来源于OCR（允许用户随时修改OCR结果并重翻）。
-				// 2. 或者，内容来源于剪贴板监听（无条件允许重翻）。
-				// 3. 或者，是纯手动输入状态 并且 开启了“输入时自动翻译”功能。   
-				//上面的不对,现在改成任意情况下,只要是双栏窗口,就可以进行编辑后自动重新翻译   
-        		// if (isContentFromOcr || (isPureManualInput && autoTranslateInputEnabled))
-				// {
-					Debug.WriteLine("        |--> 满足 [isContentFromOcr 或 autoTranslateInputEnabled]，准备启动/重置 定时器...");
-        		    translationTimer.Stop();
-                // 【修改后新增】在此处动态更新 Timer 的间隔
-                // 这样可以确保：
-                // 1. 此时 currentDelay 肯定 > 0，赋值安全，不会崩溃
-                // 2. 如果用户在设置里修改了时间，这里会立即生效，无需重启
-                // === 【修改步骤 3】使用解析出的有效时间 ===
-                translationTimer.Interval = validDelay;
-                translationTimer.Start();
-
-					Debug.WriteLine("        |--> 定时器已重置。");
-
-        		// }
-    		}
-			Debug.WriteLine("---> TextChanged 事件结束。");
-
-		}
-		
-
-		private void HandleClipboardChange()
-		{
-			Debug.WriteLine("HandleClipboardChange执行了");
-    		// 最优先检查：如果程序正处于自动复制后的静默期，则忽略所有剪贴板事件
-    		if (isAutoCopying)
-    		{
-    		    Debug.WriteLine("忽略剪贴板事件：自动复制锁激活中");
-    		    return;
-    		}
-    		
-			// 在程序刚启动时，忽略第一次自动触发的剪贴板消息
-			if (isAppLoading)
-			{
-				isAppLoading = false; // 将标志位置为false，确保此逻辑只执行一次
-                // 使用 try-catch 保护剪贴板读取，防止启动时其他程序占用剪贴板导致崩溃
-                try
-                {
-                    if (Clipboard.ContainsText())
-                    {
-                        // 同步初始的剪贴板内容，以便下一次真正的复制可以被正确比较
-                        lastClipboardText = Clipboard.GetText();
-                    }
-                }
-                catch (Exception ex)
-                {
-                    Debug.WriteLine("程序启动时同步剪贴板状态失败: " + ex.Message);
-                    // 启动时读取失败没关系，直接忽略即可，不影响后续监听
-                }
-                return; // 关键：直接退出，不执行任何后续的翻译操作
-			}
-			
-			//  检查功能是否开启
-			if (!StaticValue.ListenClipboardTranslation)
-			{
-				return;
-			}
-
-			// 不立即处理，而是重置防抖定时器
-			// 无论来多少次通知，都只是不断地重置计时器
-			clipboardDebounceTimer.Stop();
-			clipboardDebounceTimer.Start();
-			//定时器防抖解决监听剪贴板时多次翻译的问题,此问题原因是剪贴板软件获取多次系统消息导致:应注意剪贴板查看器链(不同剪贴板软件和系统的交互链路不同)
-			
-		}
-		private void ClipboardDebounceTimer_Tick(object sender, EventArgs e)
-		{
-		    // 1. 首先停止定时器，防止重复执行
-		    clipboardDebounceTimer.Stop();
-
-            string clipboardText = null;
-
-            // 2. 将 ContainsText 和 GetText 全部放入 try-catch 中
-            try
-            {
-                if (Clipboard.ContainsText())
-                {
-                    clipboardText = Clipboard.GetText();
-                }
-            }
-            catch (Exception ex)
-            {
-                Debug.WriteLine("获取剪贴板文本失败，可能被其他程序占用: " + ex.Message);
-                return; // 获取失败则直接返回，等待下一次用户复制
-            }
-
-            // 3. 检查是否是新的、非空的文本
-            if (!string.IsNullOrEmpty(clipboardText) && clipboardText != lastClipboardText)
-			{
-				// 更新最后一次的文本记录，防止重复触发
-				lastClipboardText = clipboardText;
-
-				// 显示提示
-				CommonHelper.ShowHelpMsg("已捕获剪贴板，正在翻译...");
-
-				// 设置正确的标志位
-				isContentFromOcr = false;
-				isFromClipboardListener = true;
-
-				// 调用UI启动方法
-				InitiateTranslationUI(clipboardText);
-			
-		    }
-		    
-		}
-
 #endregion
 
-// ====================================================================================================================
-		// **截图与图像处理**
-		//
-		// 包含了屏幕截图、二维码扫描、图像处理和文件操作等辅助功能。
-		// - CreateParams: 设置窗口样式，启用无边框窗口特性。
-		// - ScanQRCode(): 扫描屏幕截图中的二维码并返回解码后的文本。
-		// - SearchSelText(): 使用默认浏览器搜索选中的文本。
-		// - tray_update_Click(): 检查应用程序更新。
-		// - contain_jap(), contain_kor(): 判断字符串是否包含日文或韩文字符。
-		// - ReFileName(), GetUniqueFileName(): 生成唯一的文件名以避免覆盖。
-		// - PlaySong(): 播放音频文件。
-		// ====================================================================================================================
-#region 截图与图像处理
 		/// <summary>
 		/// 创建窗口参数，设置窗口的扩展样式
 		/// </summary>
@@ -1425,337 +854,6 @@ namespace TrOCR
 			}
 		}
 
-		/// <summary>
-		/// 扫描屏幕图像中的二维码内容
-		/// </summary>
-		/// <returns>二维码中的文本内容，如果扫描失败则返回空字符串</returns>
-		public string ScanQRCode()
-		{
-			var result = "";
-			try
-			{
-				var image = new BinaryBitmap(new HybridBinarizer(new BitmapLuminanceSource((Bitmap)image_screen)));
-				var result2 = new QRCodeReader().decode(image);
-				if (result2 != null)
-				{
-					result = result2.Text;
-				}
-			}
-			catch
-			{
-			}
-			return result;
-		}
-
-		/// <summary>
-		/// 使用百度搜索RichBoxBody控件中选中的文本
-		/// </summary>
-		/// <param name="sender">事件发送者</param>
-		/// <param name="e">事件参数</param>
-		public void SearchSelText(object sender, EventArgs e)
-		{
-			Process.Start("https://www.baidu.com/s?wd=" + RichBoxBody.SelectText);
-		}
-
-
-		/// <summary>
-		/// 重命名文件，在文件名后添加序号以避免重复
-		/// </summary>
-		/// <param name="strFolderPath">文件夹路径</param>
-		/// <param name="strFileName">原始文件名</param>
-		/// <returns>新的文件名</returns>
-		public static string ReFileName(string strFolderPath, string strFileName)
-		{
-			var text = strFolderPath + "\\" + strFileName;
-			var startIndex = text.LastIndexOf('.');
-			text = text.Insert(startIndex, "_{0}");
-			var num = 1;
-			var path = string.Format(text, num);
-			while (File.Exists(path))
-			{
-				path = string.Format(text, num);
-				num++;
-			}
-			return Path.GetFileName(path);
-		}
-
-#endregion
-
-// ====================================================================================================================
-		// **右键菜单 - 朗读事件**
-		//
-		// 处理原文和译文文本框中通过右键菜单触发的朗读功能。
-		// - Main_Voice_Click(): 获取原文框中选中的文本并触发朗读。
-		// - Trans_Voice_Click(): 获取译文框中选中的文本并触发朗读。
-		// - Speak_child(): 在 UI 线程中播放朗读音频。
-		// ====================================================================================================================
-// ====================================================================================================================
-		// **右键菜单 - 文本转换**
-		//
-		// 提供文本的大小写转换和简繁体转换功能。
-		// - change_zh_tra_Click(): 将文本转换为繁体。
-		// - change_tra_zh_Click(): 将文本转换为简体。
-		// - change_str_Upper_Click(): 将文本转换为大写。
-		// - change_Upper_str_Click(): 将文本转换为小写。
-		// ====================================================================================================================
-// ====================================================================================================================
-		// **热键管理**
-		//
-		// 负责解析快捷键字符串并注册系统范围的全局热键。
-		// - SetHotkey(): 核心方法，调用 Win32 API (RegisterHotKey) 来注册一个全局热键，
-		//              允许用户在任何地方通过快捷键触发程序功能（如截图、翻译）。
-		// ====================================================================================================================
-#region 热键管理
-
-		/// <summary>
-		/// 设置并注册全局热键
-		/// </summary>
-		/// <param name="text">修饰键（如Ctrl、Alt等）</param>
-		/// <param name="text2">按键（如A、B、F1等）</param>
-		/// <param name="value">完整的快捷键字符串，格式如"Ctrl+Alt+A"或"Alt+A"</param>
-		/// <param name="flag">热键标识符</param>
-		public void SetHotkey(string text, string text2, string value, int flag)
-		{
-			var parts = value.Split(new[] { '+' }, StringSplitOptions.RemoveEmptyEntries).Select(p => p.Trim()).ToList();
-			if (parts.Count == 0) return;
-
-			Keys key = (Keys)Enum.Parse(typeof(Keys), parts.Last());
-			HelpWin32.KeyModifiers keyModifiers = HelpWin32.KeyModifiers.None;
-
-			foreach (var part in parts.Take(parts.Count - 1))
-			{
-				if (part.Equals("Ctrl", StringComparison.OrdinalIgnoreCase))
-					keyModifiers |= HelpWin32.KeyModifiers.Ctrl;
-				else if (part.Equals("Shift", StringComparison.OrdinalIgnoreCase))
-					keyModifiers |= HelpWin32.KeyModifiers.Shift;
-				else if (part.Equals("Alt", StringComparison.OrdinalIgnoreCase))
-					keyModifiers |= HelpWin32.KeyModifiers.Alt;
-			}
-			try
-			{
-				if (!HelpWin32.RegisterHotKey(Handle, flag, keyModifiers, key))
-				{
-					CommonHelper.ShowHelpMsg($"快捷键 '{value}' 注册失败，可能已被其他程序占用！");
-
-				}
-			}
-			catch (Exception ex)
-			{
-				// 捕获任何异常，避免程序崩溃
-				CommonHelper.ShowHelpMsg($"热键注册失败: {ex.Message}");
-			}
-		}
-#endregion
-// ====================================================================================================================
-		// **辅助方法与工具**
-		//
-		// 包含一些通用的辅助方法，例如记录管理和剪贴板操作。
-		// - p_note(): 将新的识别结果添加到历史记录队列中。
-		// - GetTextFromClipboard(): 从系统剪贴板安全地获取文本内容，处理线程问题。
-		// ====================================================================================================================
-#region 辅助方法与工具
-		/// <summary>
-		/// 将指定的字符串添加到历史记录队列中
-		/// </summary>
-		/// <param name="a">要添加到历史记录的字符串</param>
-		public void p_note(string a)
-		{
-			// 循环更新历史记录数组，实现队列效果
-			for (var i = 0; i < StaticValue.NoteCount; i++)
-			{
-				if (i == StaticValue.NoteCount - 1)
-				{
-					pubnote[StaticValue.NoteCount - 1] = a;
-				}
-				else
-				{
-					pubnote[i] = pubnote[i + 1];
-				}
-			}
-		}
-
-		/// <summary>
-		/// 从系统剪贴板安全地获取文本内容
-		/// </summary>
-		/// <returns>剪贴板中的文本内容，如果为空则返回null</returns>
-		private string GetTextFromClipboard()
-		{
-			// 检查当前线程的单元状态，确保在STA模式下执行剪贴板操作
-			if (Thread.CurrentThread.GetApartmentState() > ApartmentState.STA)
-			{
-				var thread = new Thread(delegate()
-				{
-					SendKeys.SendWait("^c");
-					SendKeys.Flush();
-				});
-				thread.SetApartmentState(ApartmentState.STA);
-				thread.Start();
-				thread.Join();
-			}
-			else
-			{
-				SendKeys.SendWait("^c");
-				SendKeys.Flush();
-			}
-			var text = Clipboard.GetText();
-			text = (string.IsNullOrWhiteSpace(text) ? null : text);
-			// 如果获取到文本内容，则清空剪贴板
-			if (text != null)
-			{
-				Clipboard.Clear();
-			}
-			return text;
-		}
-#endregion
-
-// ====================================================================================================================
-		// **截图与核心OCR流程**
-		//
-		// 这是应用程序的核心功能所在，集成了截图、图像处理和OCR识别的完整流程。
-		// - MainOCRQuickScreenShots(): 启动截图功能，隐藏主窗口，调用 ShareX 库进行区域捕捉。
-		//                             根据用户的操作（如截图、贴图、保存、多区域选择等）执行不同逻辑。
-		// - Main_OCR_Thread(): 截图完成后，在此线程中执行 OCR 识别。
-		//                      它会先尝试扫描二维码，然后根据当前选择的 OCR 接口调用相应的识别方法。
-		// - Main_OCR_Thread_last(): OCR 识别完成后，在 UI 线程中更新界面，显示识别结果，处理自动翻译、
-		//                           分段合并等后续操作，并重新显示主窗口。
-		// - SougouOCR(): 调用搜狗OCR。
-		// - BdTableOCR(), OCR_ali_table(): 处理表格识别。
-		// - select_image(), FindBundingBox(): 使用 Emgu.CV进行图像处理，用于竖排文字的识别。
-		// ====================================================================================================================
-#region 截图与核心OCR流程
-
-
-		/// <summary>
-		/// 刷新ocr接口界面控件文本显示
-		/// </summary>
-		private new void Refresh()
-		{
-			sougou.Text = "搜狗";
-			tencent.Text = "腾讯";
-			tencent_accurate.Text = "腾讯-高精度";
-			baidu.Text = "百度";
-			youdao.Text = "有道";
-			wechat.Text = "微信";
-			baimiao.Text = "白描";
-			baidu_accurate.Text = "百度-高精度";
-			shupai.Text = "竖排";
-			ocr_table.Text = "表格";
-			ch_en.Text = "中英";
-			jap.Text = "日语";
-			kor.Text = "韩语";
-			left_right.Text = "从左向右";
-			righ_left.Text = "从右向左";
-			baidu_table.Text = "百度";
-			tx_table.Text = "腾讯";
-			ali_table.Text = "阿里";
-			Mathfuntion.Text = "公式";
-			paddleocr.Text = "PaddleOCR";
-			paddleocr2.Text = "PaddleOCR2";
-			rapidocr.Text = "RapidOCR";
-			write.Text = "手写";
-			baidu_handwriting.Text = "百度手写";
-			ai_menu.Text = "AI";
-
-		}
-
-
-#endregion
-		//控制右键菜单-接口项的可见性
-		private void InitializeApiMenus()
-		{
-			// OCR 接口可见性设置
-			SetMenuItemVisibility(sougou, "Ocr接口显示", "Sougou");
-			SetMenuItemVisibility(tencent, "Ocr接口显示", "Tencent");
-			SetMenuItemVisibility(tencent_accurate, "Ocr接口显示", "TencentAccurate");
-			SetMenuItemVisibility(youdao, "Ocr接口显示", "Youdao");
-			SetMenuItemVisibility(wechat, "Ocr接口显示", "WeChat");
-			SetMenuItemVisibility(baimiao, "Ocr接口显示", "Baimiao");
-			SetMenuItemVisibility(baidu, "Ocr接口显示", "Baidu");
-			SetMenuItemVisibility(baidu_accurate, "Ocr接口显示", "BaiduAccurate");
-			SetMenuItemVisibility(Mathfuntion, "Ocr接口显示", "Mathfuntion");
-			SetMenuItemVisibility(ocr_table, "Ocr接口显示", "Table");
-			SetMenuItemVisibility(shupai, "Ocr接口显示", "Shupai");
-			SetMenuItemVisibility(paddleocr, "Ocr接口显示", "PaddleOCR");
-			SetMenuItemVisibility(paddleocr2, "Ocr接口显示", "PaddleOCR2");
-			SetMenuItemVisibility(rapidocr, "Ocr接口显示", "RapidOCR");
-
-			// OCR 子菜单接口可见性设置
-			SetMenuItemVisibility(baidu_table, "Ocr接口显示", "TableBaidu");
-			SetMenuItemVisibility(tx_table, "Ocr接口显示", "TencentTable");
-			SetMenuItemVisibility(ali_table, "Ocr接口显示", "TableAli");
-			SetMenuItemVisibility(left_right, "Ocr接口显示", "ShupaiLR");
-			SetMenuItemVisibility(righ_left, "Ocr接口显示", "ShupaiRL");
-
-			// 翻译接口可见性设置
-			SetMenuItemVisibility(trans_google, "翻译接口显示", "Google");
-			SetMenuItemVisibility(trans_baidu, "翻译接口显示", "Baidu");
-			SetMenuItemVisibility(trans_tencent, "翻译接口显示", "Tencent");
-			SetMenuItemVisibility(trans_bing, "翻译接口显示", "Bing");
-			SetMenuItemVisibility(trans_bing2, "翻译接口显示", "Bing2");
-			SetMenuItemVisibility(trans_microsoft, "翻译接口显示", "Microsoft");
-			SetMenuItemVisibility(trans_yandex, "翻译接口显示", "Yandex");
-			SetMenuItemVisibility(trans_tencentinteractive, "翻译接口显示", "TencentInteractive");
-			SetMenuItemVisibility(trans_caiyun, "翻译接口显示", "Caiyun");
-			SetMenuItemVisibility(trans_volcano, "翻译接口显示", "Volcano");
-			SetMenuItemVisibility(trans_caiyun2, "翻译接口显示", "Caiyun2");
-			SetMenuItemVisibility(trans_baidu2, "翻译接口显示", "Baidu2");
-		}
-
-		/// <summary>
-		/// 根据配置文件中指定节和键的值设置菜单项的可见性
-		/// </summary>
-		/// <param name="menuItem">要设置可见性的菜单项</param>
-		/// <param name="section">配置文件中的节名称</param>
-		/// <param name="key">节中的键名称</param>
-		private void SetMenuItemVisibility(ToolStripItem menuItem, string section, string key)
-		{
-			if (menuItem != null)
-			{
-				string visibilityValue = IniHelper.GetValue(section, key);
-				bool isVisible;
-
-				if (bool.TryParse(visibilityValue, out isVisible))
-				{
-					// Value was "True" or "False", isVisible is now set correctly.
-				}
-				else // Value was "发生错误" or something else. Apply default logic.
-				{
-					if (section == "翻译接口显示")
-					{
-						switch (key)
-						{
-							case "TencentInteractive":
-							case "Caiyun":
-							case "Volcano":
-							case "Baidu2":
-								isVisible = false;
-								break;
-							default:
-								isVisible = true;
-								break;
-						}
-					}
-					else if (section == "Ocr接口显示")
-					{
-						switch (key)
-						{
-							case "Baimiao":
-								isVisible = false;
-								break;
-							default:
-								isVisible = true;
-								break;
-						}
-					}
-					else
-					{
-						isVisible = true; // Default for any other section
-					}
-				}
-				
-				menuItem.Visible = isVisible;
-			}
-		}
 
 // ====================================================================================================================
 		// **字段声明**
@@ -2288,23 +1386,6 @@ namespace TrOCR
 
 				private bool _preformatted;
 			}
-		}
-
-		// 计算MD5值
-		public static string EncryptString(string str)
-		{
-			var md5 = MD5.Create();
-			// 将字符串转换成字节数组
-			var byteOld = Encoding.UTF8.GetBytes(str);
-			// 调用加密方法
-			var byteNew = md5.ComputeHash(byteOld);
-			// 将加密结果转换为字符串
-			var sb = new StringBuilder();
-			foreach (var b in byteNew)
-				// 将字节转换成16进制表示的字符串，
-				sb.Append(b.ToString("x2"));
-			// 返回加密的字符串
-			return sb.ToString();
 		}
